rename widget-header-size to widget-header-fit
the file holds the fit contexts, not just sizing
