+++ packages/widget-dashboard/src/components/widget-header/widget-header-fit.ts
